pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "iqp-mmd"
version = "0.1.0"
description = "Modular toolkit for IQP circuit training with MMD loss — based on XanaduAI/scaling-gqml"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
dependencies = [
    "jax>=0.4.20",
    "jaxlib>=0.4.20",
    "flax>=0.8.0",
    "numpy>=1.24",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "matplotlib>=3.7",
    "seaborn>=0.13",
    "pyyaml>=6.0",
    "networkx>=3.0",
    "iqpopt>=2024.7.0",
    "pennylane>=0.33",
    "numpyro>=0.13",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.1"]
datasets = ["torch>=2.0", "torchvision>=0.15", "requests>=2.28", "joblib>=1.3"]
benchmarks = ["qml-benchmarks>=0.1"]

[project.scripts]
iqp-train = "iqp_mmd.cli.train:main"
iqp-evaluate = "iqp_mmd.cli.evaluate:main"
iqp-sample = "iqp_mmd.cli.sample:main"
iqp-dataset = "iqp_mmd.cli.generate_dataset:main"

[tool.hatch.build.targets.wheel]
packages = ["src/iqp_mmd"]

[tool.ruff]
line-length = 120
target-version = "py310"
